+++ src/core/class/index.ts
@@ -4,3 +4,11 @@
 
 export * from './types';
 export { ClassBuilder } from './ClassBuilder';
+export {
+  runClass,
+  startClassUnitTestRun,
+  getClassUnitTestStatus,
+  getClassUnitTestResult,
+  ClassUnitTestDefinition,
+  ClassUnitTestRunOptions
+} from './run';